settings: replace console with log4js calls
log4js.replaceConsole was removed from log4js

this implementation does a similar thing to replace the methods on the global
console object.

Replacing console calls in server.js is not redone, as it would not achieve the
intended goal. Replacing the calls like this will only work after the log4js
configuration was loaded.

If it is replaced in server.js before the configuration is loaded it would use
the default configuration until the replacement in settings.js has happened. The
log4js default configuration is by default not logging anything, so some startup
messages would get lost unintentionally if console replacement happened too
early.