settings: provide socket.io loglevel from log4js
before we used the setting `loglevel` but that is not used anymore, instead
we load a logger for `socket.io` category and take its loglevel. If nothing,
special is configured this is the same as the default loglevel.

But it allows to have a separate loglevel for socket io by specifying a log4js
category `socket.io` in the `log4js_configuration` settings.